requirements.txt
flask==3.0.2
gunicorn==21.2.0
psycopg2-binary==2.9.9
cryptography==42.0.5
pandas==2.2.1
plotly==5.19.0
ccxt==4.2.14
python-dotenv==1.0.1